+++ src/main/java/com/moneygest/Repository/IngresoRepository.java
@@ -1,11 +1,12 @@
 package com.moneygest.Repository;
 
-import com.moneygest.model.Ingreso;
+import com.moneygest.Model.Ingreso;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface IngresoRepository extends JpaRepository<Ingreso, Long> {
+public interface IngresoRepository extends JpaRepository<Ingreso, Integer> {
 
     @Query("SELECT COALESCE(SUM(i.monto), 0) FROM Ingreso i")
     double totalIngresos();
+
 }
